Reorganize into spelling module
