+++ proselint/.proselintrc
@@ -21,7 +21,6 @@
         "garner.jargon"                 : true,
         "garner.malapropisms"           : true,
         "garner.many_a"                 : true,
-        "garner.misspelling"            : true,
         "garner.mixed_metaphors"        : true,
         "garner.mondegreens"            : true,
         "garner.needless_variants"      : true,
@@ -61,17 +60,24 @@
         "pinker.metaconcepts"           : true,
         "pinker.narcisissm"             : true,
         "pinker.scare_quotes"           : true,
-        "szekely.nword"                 : true,
+        "skunked_terms.misc"            : true,
+        "spelling.athletes"             : true,
+        "spelling.misc"                 : true,
+        "spelling.able_atable"          : true,
+        "spelling.able_ible"            : true,
+        "spelling.em_im_en_in"          : true,
+        "spelling.er_or"                : true,
+        "spelling.in_un"                : true,
         "strunk_white.composition"      : true,
         "strunk_white.greylist"         : true,
         "strunk_white.usage"            : true,
+        "szekely.nword"                 : true,
         "typography.symbols"            : true,
         "wallace.redundancy"            : true,
         "wallace.tense_present"         : true,
         "wallace.uncomparables"         : true,
         "white.damn"                    : true,
         "write_good.lexical_illusions"  : true,
-        "write_good.weasel_words"       : true,
-        "wsj.athletes"                  : true
+        "write_good.weasel_words"       : true
     }
 }
