--- proselint/checks/wsj/__init__.py
@@ -1,1 +1,0 @@
-"""Advice from The Wall Street Journal."""
